# Conflicts:
#	Assets/Resources/Throwing/ThrowingObject.prefab
#	Assets/Scenes/Prod/Level1.unity
#	ProjectSettings/ProjectSettings.asset

+++ Assets/Scripts/Menu&UI/MainMenu.cs
@@ -33,7 +33,7 @@ public class MainMenu : MonoBehaviour
 		{
 			PlayerPrefs.SetString ("Throw", "ClassicThrow");
 			PlayerPrefs.SetFloat ("ThrowAttackStat", 1);
-			PlayerPrefs.SetInt ("ThrowSpeedStat", 14);
+			PlayerPrefs.SetFloat ("ThrowSpeedStat", 14);
 		}
